refactor: move the internal-address check from webhooks to core
