+++ api/core/network.py
@@ -1,0 +1,26 @@
+import ipaddress
+import socket
+
+
+def is_internal_address(hostname: str) -> bool:
+    """Return True if the hostname is, or resolves to, an internal network
+    address: loopback, RFC 1918 private, link-local, reserved, or multicast.
+    Unresolvable hostnames are not considered internal."""
+    try:
+        ips = [ipaddress.ip_address(hostname)]
+    except ValueError:
+        # hostname is a name rather than a literal IP — resolve it.
+        try:
+            results = socket.getaddrinfo(hostname, None, socket.AF_UNSPEC)
+            ips = [ipaddress.ip_address(str(r[4][0]).split("%")[0]) for r in results]
+        except socket.gaierror:
+            return False
+
+    return any(
+        ip.is_loopback
+        or ip.is_private
+        or ip.is_link_local
+        or ip.is_reserved
+        or ip.is_multicast
+        for ip in ips
+    )
